# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 캠퍼스라이프 API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface Schedule {
  id: number;
  subjectName: string;
  professor?: string;
  location?: string;
  /** 0=월, 1=화, 2=수, 3=목, 4=금, 5=토 */
  dayOfWeek: number;
  /** HH:MM 형식 (예: 09:00) */
  startTime: string;
  /** HH:MM 형식 (예: 10:30) */
  endTime: string;
  color: string;
  createdAt: string;
}

export interface CreateScheduleRequest {
  subjectName: string;
  professor?: string;
  location?: string;
  dayOfWeek: number;
  startTime: string;
  endTime: string;
  color: string;
}

export type FinanceType = (typeof FinanceType)[keyof typeof FinanceType];

export const FinanceType = {
  income: "income",
  expense: "expense",
} as const;

export interface Finance {
  id: number;
  type: FinanceType;
  amount: number;
  category: string;
  description?: string;
  date: string;
  createdAt: string;
}

export type CreateFinanceRequestType =
  (typeof CreateFinanceRequestType)[keyof typeof CreateFinanceRequestType];

export const CreateFinanceRequestType = {
  income: "income",
  expense: "expense",
} as const;

export interface CreateFinanceRequest {
  type: CreateFinanceRequestType;
  amount: number;
  category: string;
  description?: string;
  date: string;
}

export interface FinanceSummary {
  totalIncome: number;
  totalExpense: number;
  balance: number;
  month: string;
}

export interface Restaurant {
  id: number;
  name: string;
  category: string;
  address?: string;
  phone?: string;
  description?: string;
  /** 학생회 제휴 여부 */
  isAffiliated: boolean;
  discountInfo?: string;
  averageRating: number;
  reviewCount: number;
  imageUrl?: string;
  createdAt: string;
}

export interface CreateRestaurantRequest {
  name: string;
  category: string;
  address?: string;
  phone?: string;
  description?: string;
  isAffiliated: boolean;
  discountInfo?: string;
  imageUrl?: string;
}

export interface Review {
  id: number;
  restaurantId: number;
  nickname: string;
  /** 1-5점 */
  rating: number;
  content: string;
  visitDate?: string;
  createdAt: string;
}

export interface CreateReviewRequest {
  nickname: string;
  rating: number;
  content: string;
  visitDate?: string;
}

export type GetFinancesParams = {
  /**
   * YYYY-MM 형식
   */
  month?: string;
};

export type GetFinanceSummaryParams = {
  /**
   * YYYY-MM 형식
   */
  month?: string;
};

export type GetRestaurantsParams = {
  category?: string;
  /**
   * 학생회 제휴 업체만 필터링
   */
  affiliated?: boolean;
};