# lib/api-zod/src/generated/types/createReviewRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 캠퍼스라이프 API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateReviewRequest {
  nickname: string;
  rating: number;
  content: string;
  visitDate?: string;
}