# lib/api-zod/src/generated/types/createScheduleRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 캠퍼스라이프 API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateScheduleRequest {
  subjectName: string;
  professor?: string;
  location?: string;
  dayOfWeek: number;
  startTime: string;
  endTime: string;
  color: string;
}