# lib/api-zod/src/generated/types/finance.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 캠퍼스라이프 API
 * OpenAPI spec version: 0.1.0
 */
import type { FinanceType } from "./financeType";

export interface Finance {
  id: number;
  type: FinanceType;
  amount: number;
  category: string;
  description?: string;
  date: string;
  createdAt: string;
}