# lib/api-zod/src/generated/types/restaurant.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 캠퍼스라이프 API
 * OpenAPI spec version: 0.1.0
 */

export interface Restaurant {
  id: number;
  name: string;
  category: string;
  address?: string;
  phone?: string;
  description?: string;
  /** 학생회 제휴 여부 */
  isAffiliated: boolean;
  discountInfo?: string;
  averageRating: number;
  reviewCount: number;
  imageUrl?: string;
  createdAt: string;
}