# lib/api-zod/src/generated/types/review.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 캠퍼스라이프 API
 * OpenAPI spec version: 0.1.0
 */

export interface Review {
  id: number;
  restaurantId: number;
  nickname: string;
  /** 1-5점 */
  rating: number;
  content: string;
  visitDate?: string;
  createdAt: string;
}