# lib/api-zod/src/generated/types/schedule.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 캠퍼스라이프 API
 * OpenAPI spec version: 0.1.0
 */

export interface Schedule {
  id: number;
  subjectName: string;
  professor?: string;
  location?: string;
  /** 0=월, 1=화, 2=수, 3=목, 4=금, 5=토 */
  dayOfWeek: number;
  /** HH:MM 형식 (예: 09:00) */
  startTime: string;
  /** HH:MM 형식 (예: 10:30) */
  endTime: string;
  color: string;
  createdAt: string;
}